[FEAT] 주소 연동 및, 제스처 + 터치 이벤트 연동
