+++ Atcha-iOS/Presentation/Location/View/LastTrainSearchBottomView.swift
@@ -114,6 +114,11 @@ final class LastTrainSearchBottomView: UIView {
             make.top.equalTo(arrivalLocationView.snp.bottom).inset(-24)
         }
     }
+    
+    func setupCurrentLocationTitle(_ address: String) {
+        // TODO: - 현재 내 위치랑 동일한 경우, 현위치 포함 아닌경우 pass
+        currentLocationLabel.attributedText = AtchaFont.B1_R_17("현위치 : \(address)", color: .main)
+    }
 }
 
 extension LastTrainSearchBottomView {
